Movie_Rental/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
jose
passlib